receiver/requirements.txt
fastapi==0.75.2
kafka-python==2.0.2